refactor: `ConsoleApp` 클래스 타입 파일 추가
